integer/numeric -> double

+++ R/check_claim.R
@@ -6,6 +6,6 @@
 #' @keywords claim boolean matrix
 check_claim <- function(code, target) {
   vapply(code,
-         function(x) as.integer(grepl(x, target, perl = TRUE)),
-         FUN.VALUE = integer(length(target)), USE.NAMES = FALSE)
+         function(x) as.double(grepl(x, target, perl = TRUE)),
+         FUN.VALUE = double(length(target)), USE.NAMES = FALSE)
 }
